# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI-Астростилист Backend API
  description: Backend API для Telegram-бота. Профиль пользователя, генерация рекомендаций, каталог товаров и правила маппинга.
  version: 2.0.0
servers:
  - url: https://api.astrostylist.example.com
    description: Production
  - url: http://localhost:8080
    description: Local
security:
  - botAuth: []
paths:
...
  /api/v1/admin/rules:
    get:
      summary: Список правил маппинга
      tags: [Admin — Rules]
      security:
        - adminAuth: []
      parameters:
        - name: is_active
          in: query
          schema:
            type: boolean
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: page
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        "200":
          description: Список правил с пагинацией
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRulesListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerErr"

    post:
      summary: Создать правило маппинга
      tags: [Admin — Rules]
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRulePayload"
            example:
              name: "Венера в Тельце"
              astro_condition:
                planet: "Venus"
                sign: "Taurus"
              product_tags: ["luxury", "comfort", "classic"]
              priority: 10
              is_active: true
      responses:
        "201":
          description: Правило создано
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerErr"
  /api/v1/admin/rules/{id}:
    get:
      summary: Получение правила по id
      tags: [Admin — Rules]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated astro rules list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRulesSingleEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerErr"
    put:
      summary: Полное обновление правила
      tags: [Admin — Rules]
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRulePayload"
      responses:
        "200":
          description: Правило обновлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerErr"
    patch:
      summary: Частичное обновление правила
      tags: [Admin — Rules]
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRulePayload"
      responses:
        "200":
          description: Правило обновлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerErr"

    delete:
      summary: Деактивировать правило (soft delete)
      tags: [Admin — Rules]
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Правило деактивировано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminRuleEnvelope"

        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErr"
...
  /api/v1/admin/logs:
    get:
      summary: Лог генераций (requests_log)
      tags: [Admin — Monitoring]
      security:
        - adminAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Лог запросов с пагинацией
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLogsListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    botAuth:
      type: http
      scheme: bearer
      description: Bearer-токен для Telegram-бота
    adminAuth:
      type: http
      scheme: bearer
      description: ADMIN_TOKEN из переменных окружения
  responses:
    BadRequest:
      description: Ошибка валидации запроса
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Unauthorized:
      description: Отсутствует или невалидный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    InternalServerErr:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              error:
                type: string
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              error:
                type: string
    ValidationError:
      description: Ошибка бизнес-валидации
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
...
  schemas:
    Envelope:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        data:
          nullable: true
...
    AdminRulePayload:
      type: object
      required: [name, astro_condition]
      properties:
        name:
          type: string
          maxLength: 255
        astro_condition:
          type: object
          additionalProperties: true
          description: Пример — {"planet":"Venus","sign":"Taurus"}
        product_tags:
          type: array
          items:
            type: string
        priority:
          type: integer
          minimum: 0
          default: 99
        is_active:
          type: boolean
          default: true
    AdminRule:
      type: object
      required:
        [
          id,
          name,
          astro_condition,
          product_tags,
          priority,
          is_active,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        astro_condition:
          type: object
          additionalProperties: true
        product_tags:
          type: array
          items:
            type: string
        priority:
          type: integer
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminRuleEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/AdminRule"
    AdminRulesListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/AdminRule"
                metadata:
                  type: object
                  properties:
                    current_page:
                      type: integer
                    page_size:
                      type: integer
                    first_page:
                      type: integer
                    last_page:
                      type: integer
                    total_records:
                      type: integer
    AdminRulesSingleEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                rule:
                  $ref: "#/components/schemas/AdminRule"
...
    AdminLogEntry:
      type: object
      required:
        - request_id
        - user_id
        - status
        - error_message
        - created_at
        - completed_at
      properties:
        request_id:
          type: string
          format: uuid
        user_id:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        error_message:
          type: string
          maxLength: 500
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
...
    AdminLogsListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/AdminLogEntry"
                total_count:
                  type: integer